Environment improved (#19)
This pull request includes several changes to the `Dockerfile` and the
addition of new shell scripts to improve the setup process for
development environments. The most important changes include the removal
of redundant user creation steps, the addition of locale setup, and the
introduction of new shell scripts for user creation and locale setup.

### Dockerfile Improvements:

* Removed redundant user creation steps and consolidated them into a new
shell script `create_user.sh`.
* Added locale setup by including a new shell script `setup_locale.sh`.

### New Shell Scripts:

*
[`shells/root/create_user.sh`](diffhunk://#diff-b9b714d64391d253fd4ac8c14212d1aa2b1e2a005157aded9972af1b7bd926a1R1-R15):
A new script to remove an unused user and create a non-root user.
*
[`shells/root/setup_locale.sh`](diffhunk://#diff-3f6e8a75b6ec3cd50263c68a01d4a8b4d128d80632a98d0632cfdc2742295750R1-R7):
A new script to configure the system locale based on the provided
argument.

### Other Changes:

* Updated the `Dockerfile` to use wildcard `*.sh` for setting executable
permissions on shell scripts.